+++ test/input-range/input-range.spec.jsx
@@ -2,36 +2,35 @@ import React from 'react';
 import InputRange from '../../src/js';
 import { mount, shallow } from 'enzyme';
 
-let container;
-let requestAnimationFrame;
-
-beforeEach(() => {
-  requestAnimationFrame = window.requestAnimationFrame;
-  window.requestAnimationFrame = callback => callback();
-
-  container = document.createElement('div');
-  document.body.appendChild(container);
-});
-
-afterEach(() => {
-  window.requestAnimationFrame = requestAnimationFrame;
-
-  document.body.removeChild(container);
-});
-
 describe('InputRange', () => {
+  let container;
+  let requestAnimationFrame;
+
+  beforeEach(() => {
+    requestAnimationFrame = window.requestAnimationFrame;
+    window.requestAnimationFrame = callback => callback();
+
+    container = document.createElement('div');
+    document.body.appendChild(container);
+  });
+
+  afterEach(() => {
+    window.requestAnimationFrame = requestAnimationFrame;
+
+    document.body.removeChild(container);
+  });
+
   it('updates the current value when the user tries to drag the slider', () => {
     const jsx = (
       <InputRange
         maxValue={20}
         minValue={0}
         value={{ min: 2, max: 10 }}
-        onChange={value => component.setProps({ value })}
-      />
+        onChange={value => component.setProps({ value })} />
     );
     const component = mount(jsx, { attachTo: container });
-    const minSlider = component.find(`Slider [onMouseDown]`).at(0);
-    const maxSlider = component.find(`Slider [onMouseDown]`).at(1);
+    const minSlider = component.find('Slider [onMouseDown]').at(0);
+    const maxSlider = component.find('Slider [onMouseDown]').at(1);
 
     minSlider.simulate('mouseDown', { clientX: 50, clientY: 50 });
     document.dispatchEvent(new MouseEvent('mousemove', { clientX: 100, clientY: 50 }));
@@ -52,11 +51,10 @@ describe('InputRange', () => {
         maxValue={20}
         minValue={0}
         value={{ min: 2, max: 10 }}
-        onChange={value => component.setProps({ value })}
-      />
+        onChange={value => component.setProps({ value })} />
     );
     const component = mount(jsx, { attachTo: container });
-    const track = component.find(`Track [onMouseDown]`).first();
+    const track = component.find('Track [onMouseDown]').first();
 
     track.simulate('mouseDown', { clientX: 150, clientY: 50 });
     expect(component.props().value).toEqual({ min: 2, max: 7 });
@@ -73,11 +71,10 @@ describe('InputRange', () => {
         maxValue={20}
         minValue={0}
         value={{ min: 2, max: 10 }}
-        onChange={value => component.setProps({ value })}
-      />
+        onChange={value => component.setProps({ value })} />
     );
     const component = mount(jsx, { attachTo: container });
-    const track = component.find(`Track [onTouchStart]`).first();
+    const track = component.find('Track [onTouchStart]').first();
 
     track.simulate('touchStart', { touches: [{ clientX: 150, clientY: 50 }] });
     expect(component.props().value).toEqual({ min: 2, max: 7 });
@@ -92,11 +89,10 @@ describe('InputRange', () => {
         minValue={0}
         value={{ min: 2, max: 10 }}
         onChange={value => component.setProps({ value })}
-        step={2}
-      />
+        step={2} />
     );
     const component = mount(jsx, { attachTo: container });
-    const slider = component.find(`Slider [onMouseDown]`).first();
+    const slider = component.find('Slider [onMouseDown]').first();
 
     slider.simulate('mouseDown', { clientX: 50, clientY: 50 });
     document.dispatchEvent(new MouseEvent('mousemove', { clientX: 60, clientY: 50 }));
@@ -117,11 +113,10 @@ describe('InputRange', () => {
         maxValue={20}
         minValue={0}
         value={{ min: 2, max: 10 }}
-        onChange={value => component.setProps({ value })}
-      />
+        onChange={value => component.setProps({ value })} />
     );
     const component = mount(jsx);
-    const slider = component.find(`Slider [onKeyDown]`).first();
+    const slider = component.find('Slider [onKeyDown]').first();
 
     slider.simulate('keyDown', { keyCode: 37 });
     slider.simulate('keyUp', { keyCode: 37 });
@@ -138,11 +133,10 @@ describe('InputRange', () => {
         maxValue={20}
         minValue={0}
         value={{ min: 2, max: 10 }}
-        onChange={value => component.setProps({ value })}
-      />
+        onChange={value => component.setProps({ value })} />
     );
     const component = mount(jsx);
-    const slider = component.find(`Slider [onKeyDown]`).first();
+    const slider = component.find('Slider [onKeyDown]').first();
 
     slider.simulate('keyDown', { keyCode: 65 });
     slider.simulate('keyUp', { keyCode: 65 });
@@ -152,13 +146,12 @@ describe('InputRange', () => {
   it('does not respond to mouse event when it is disabled', () => {
     const jsx = (
       <InputRange
-        disabled={true}
+        disabled
         value={{ min: 2, max: 10 }}
-        onChange={value => component.setProps({ value })}
-      />
+        onChange={value => component.setProps({ value })} />
     );
     const component = mount(jsx, { attachTo: container });
-    const slider = component.find(`Slider [onMouseDown]`).at(0);
+    const slider = component.find('Slider [onMouseDown]').at(0);
 
     slider.simulate('mouseDown', { clientX: 50, clientY: 50 });
     document.dispatchEvent(new MouseEvent('mousemove', { clientX: 100, clientY: 50 }));
@@ -171,13 +164,12 @@ describe('InputRange', () => {
   it('does not respond to keyboard event when it is disabled', () => {
     const jsx = (
       <InputRange
-        disabled={true}
+        disabled
         value={2}
-        onChange={value => component.setProps({ value })}
-      />
+        onChange={value => component.setProps({ value })} />
     );
     const component = mount(jsx);
-    const slider = component.find(`Slider [onKeyDown]`).first();
+    const slider = component.find('Slider [onKeyDown]').first();
 
     slider.simulate('keyDown', { keyCode: 37 });
     slider.simulate('keyUp', { keyCode: 37 });
@@ -190,12 +182,11 @@ describe('InputRange', () => {
         maxValue={20}
         minValue={0}
         value={{ min: 2, max: 10 }}
-        onChange={value => component.setProps({ value })}
-      />
+        onChange={value => component.setProps({ value })} />
     );
     const component = mount(jsx, { attachTo: container });
-    const minSlider = component.find(`Slider [onMouseDown]`).at(0);
-    const maxSlider = component.find(`Slider [onMouseDown]`).at(1);
+    const minSlider = component.find('Slider [onMouseDown]').at(0);
+    const maxSlider = component.find('Slider [onMouseDown]').at(1);
 
     minSlider.simulate('mouseDown', { clientX: 50, clientY: 50 });
     document.dispatchEvent(new MouseEvent('mousemove', { clientX: -20, clientY: 50 }));
@@ -216,11 +207,10 @@ describe('InputRange', () => {
         maxValue={20}
         minValue={0}
         value={2}
-        onChange={value => component.setProps({ value })}
-      />
+        onChange={value => component.setProps({ value })} />
     );
     const component = mount(jsx, { attachTo: container });
-    const slider = component.find(`Slider [onMouseDown]`).first();
+    const slider = component.find('Slider [onMouseDown]').first();
 
     slider.simulate('mouseDown', { clientX: 50, clientY: 50 });
     document.dispatchEvent(new MouseEvent('mousemove', { clientX: -20, clientY: 50 }));
@@ -241,11 +231,10 @@ describe('InputRange', () => {
         maxValue={20}
         minValue={0}
         value={{ min: 2, max: 10 }}
-        onChange={value => component.setProps({ value })}
-      />
+        onChange={value => component.setProps({ value })} />
     );
     const component = mount(jsx, { attachTo: container });
-    const slider = component.find(`Slider [onMouseDown]`).first();
+    const slider = component.find('Slider [onMouseDown]').first();
 
     slider.simulate('mouseDown', { clientX: 50, clientY: 50 });
     document.dispatchEvent(new MouseEvent('mousemove', { clientX: 190, clientY: 50 }));
@@ -255,26 +244,66 @@ describe('InputRange', () => {
     component.detach();
   });
 
+  it('allows the min value to equal the max value', () => {
+    const jsx = (
+      <InputRange
+        allowSameValues
+        maxValue={20}
+        minValue={0}
+        value={{ min: 2, max: 10 }}
+        onChange={value => component.setProps({ value })} />
+    );
+    const component = mount(jsx, { attachTo: container });
+    const minSlider = component.find('Slider [onMouseDown]').at(0);
+
+    minSlider.simulate('mouseDown', { clientX: 50, clientY: 50 });
+    document.dispatchEvent(new MouseEvent('mousemove', { clientX: 200, clientY: 50 }));
+    document.dispatchEvent(new MouseEvent('mouseup', { clientX: 200, clientY: 50 }));
+
+    expect(component.props().value).toEqual({ min: 10, max: 10 });
+
+    component.detach();
+  });
+
+  it('does not allow the min value to equal the max value', () => {
+    const jsx = (
+      <InputRange
+        maxValue={20}
+        minValue={0}
+        value={{ min: 2, max: 10 }}
+        onChange={value => component.setProps({ value })} />
+    );
+    const component = mount(jsx, { attachTo: container });
+    const minSlider = component.find('Slider [onMouseDown]').at(0);
+
+    minSlider.simulate('mouseDown', { clientX: 50, clientY: 50 });
+    document.dispatchEvent(new MouseEvent('mousemove', { clientX: 200, clientY: 50 }));
+    document.dispatchEvent(new MouseEvent('mouseup', { clientX: 200, clientY: 50 }));
+
+    expect(component.props().value).toEqual({ min: 2, max: 10 });
+
+    component.detach();
+  });
+
   it('notifies the parent component when dragging starts', () => {
     const onChange = jasmine.createSpy('onChange').and.callFake(value => component.setProps({ value }));
     const onChangeStart = jasmine.createSpy('onChangeStart');
     const jsx = (
-  	<InputRange
-  	  maxValue={20}
-  	  minValue={0}
-  	  value={{ min: 2, max: 10 }}
-  	  onChange={onChange}
-  	  onChangeStart={onChangeStart} />
-  	);
+      <InputRange
+        maxValue={20}
+        minValue={0}
+        value={{ min: 2, max: 10 }}
+        onChange={value => component.setProps({ value })}
+        onChangeStart={onChangeStart} />
+    );
     const component = mount(jsx, { attachTo: container });
-    // eslint-disable-next-line quotes
-    const slider = component.find(`Slider [onMouseDown]`).first();
+    const slider = component.find('Slider [onMouseDown]').first();
 
     slider.simulate('mouseDown', { clientX: 50, clientY: 50 });
     document.dispatchEvent(new MouseEvent('mousemove', { clientX: 100, clientY: 50 }));
     document.dispatchEvent(new MouseEvent('mousemove', { clientX: 150, clientY: 50 }));
     document.dispatchEvent(new MouseEvent('mouseup', { clientX: 150, clientY: 50 }));
-    expect(onChangeStart.calls.count()).toEqual(0);
+    expect(onChangeStart.calls.count()).toEqual(1);
 
     component.detach();
   });
@@ -288,11 +317,10 @@ describe('InputRange', () => {
         minValue={0}
         value={{ min: 2, max: 10 }}
         onChange={onChange}
-        onChangeComplete={onChangeComplete}
-      />
+        onChangeComplete={onChangeComplete} />
     );
     const component = mount(jsx, { attachTo: container });
-    const slider = component.find(`Slider [onMouseDown]`).first();
+    const slider = component.find('Slider [onMouseDown]').first();
 
     slider.simulate('mouseDown', { clientX: 50, clientY: 50 });
     document.dispatchEvent(new MouseEvent('mousemove', { clientX: 100, clientY: 50 }));
@@ -313,11 +341,10 @@ describe('InputRange', () => {
         minValue={0}
         value={{ min: 2, max: 10 }}
         onChange={onChange}
-        onChangeComplete={onChangeComplete}
-      />
+        onChangeComplete={onChangeComplete} />
     );
     const component = mount(jsx, { attachTo: container });
-    const slider = component.find(`Slider [onMouseDown]`).first();
+    const slider = component.find('Slider [onMouseDown]').first();
 
     slider.simulate('mouseDown', { clientX: 50, clientY: 50 });
     document.dispatchEvent(new MouseEvent('mousemove', { clientX: 51, clientY: 50 }));
@@ -332,8 +359,7 @@ describe('InputRange', () => {
     const jsx = (
       <InputRange
         value={{ min: 2, max: 10 }}
-        onChange={value => component.setProps({ value })}
-      />
+        onChange={value => component.setProps({ value })} />
     );
     const component = mount(jsx);
     const label = component.find('Slider Label').first();
@@ -345,9 +371,8 @@ describe('InputRange', () => {
     const jsx = (
       <InputRange
         value={{ min: 2, max: 10 }}
-        formatLabel={(value) => `${value}cm`}
-        onChange={value => component.setProps({ value })}
-      />
+        formatLabel={value => `${value}cm`}
+        onChange={value => component.setProps({ value })} />
     );
     const component = mount(jsx);
     const label = component.find('Slider Label').first();
@@ -359,8 +384,7 @@ describe('InputRange', () => {
     const jsx = (
       <InputRange
         value={{ min: 2, max: 10 }}
-        onChange={value => component.setProps({ value })}
-      />
+        onChange={value => component.setProps({ value })} />
     );
     const component = mount(jsx);
     const sliderHandle = component.find('Slider [role="slider"]').first();
@@ -372,8 +396,7 @@ describe('InputRange', () => {
     const jsx = (
       <InputRange
         value={{ min: 2, max: 10 }}
-        onChange={() => {}}
-      />
+        onChange={() => {}} />
     );
     const component = mount(jsx);
 
@@ -384,8 +407,7 @@ describe('InputRange', () => {
     const jsx = (
       <InputRange
         value={2}
-        onChange={() => {}}
-      />
+        onChange={() => {}} />
     );
     const component = mount(jsx);
 
@@ -397,8 +419,7 @@ describe('InputRange', () => {
       <InputRange
         name="price"
         value={{ min: 2, max: 10 }}
-        onChange={() => {}}
-      />
+        onChange={() => {}} />
     );
     const component = mount(jsx);
     const minInput = component.find('[name="priceMin"][type="hidden"]');
@@ -413,8 +434,7 @@ describe('InputRange', () => {
       <InputRange
         name="price"
         value={5}
-        onChange={() => {}}
-      />
+        onChange={() => {}} />
     );
     const component = mount(jsx);
     const hiddenInput = component.find('[name="price"][type="hidden"]');
@@ -428,7 +448,7 @@ describe('InputRange', () => {
       { minValue: 10, maxValue: 2 },
     ];
 
-    sampleProps.forEach(props => {
+    sampleProps.forEach((props) => {
       expect(InputRange.propTypes.minValue(props)).toEqual(jasmine.any(Error));
       expect(InputRange.propTypes.maxValue(props)).toEqual(jasmine.any(Error));
     });
@@ -443,7 +463,7 @@ describe('InputRange', () => {
       { value: null, minValue: 2, maxValue: 10 },
     ];
 
-    sampleProps.forEach(props => {
+    sampleProps.forEach((props) => {
       expect(InputRange.propTypes.value(props, 'value')).toEqual(jasmine.any(Error));
     });
   });